Rename `demangle_cw.py` --> `demangle_codewarrior.py`
